Reorganized project structure for better scalability and maintainability.
