+++ src/components/layout/HamburgerMenu.jsx
@@ -1,6 +1,6 @@
 import React, {useState} from 'react';
 import {useTranslation} from 'react-i18next';
-import LanguageSwitcher from './LanguageSwitcher';
+import LanguageSwitcher from './LanguageSwitcher.jsx';
 
 const HamburgerMenu = () => {
     const {t} = useTranslation();
