+++ src/components/sections/Projects.jsx
@@ -1,7 +1,7 @@
 // src/components/Projects.jsx
 import React from 'react';
-import {loadImage} from '../utils/loadImage';
-import ProjectCard from './ProjectCard';
+import {loadImage} from '../../utils/loadImage.js';
+import ProjectCard from '../common/ProjectCard.jsx';
 
 const Projects = () => {
     const projects = [
