+++ src/i18n/i18n.js
@@ -1,8 +1,8 @@
 import i18n from 'i18next';
 import {initReactI18next} from 'react-i18next';
-import enTranslation from './locales/en.json';
-import esTranslation from './locales/es.json';
-import caTranslation from './locales/ca.json';
+import enTranslation from '../locales/en.json';
+import esTranslation from '../locales/es.json';
+import caTranslation from '../locales/ca.json';
 
 i18n.use(initReactI18next).init({
     resources: {
