+++ src/main.jsx
@@ -2,7 +2,7 @@ import {StrictMode} from 'react'
 import {createRoot} from 'react-dom/client'
 import App from './App.jsx'
 import './assets/styles/index.css'
-import './i18n.js'
+import './i18n/i18n.js'
 
 createRoot(document.getElementById('root')).render(
     <StrictMode>
